example-rag/requirements.txt
llama-index-storage-docstore-redis
llama-index-vector-stores-redis
llama-index-embeddings-huggingface
llama-index-readers-json